physical-ai-and-humanoid-robotics-textbook/backend/rag_chatbot/requirements.txt
fastapi
pydantic
python-dotenv
asyncpg
qdrant-client
uvicorn
google-generativeai
fastembed